pyproject.toml
[project]
name = "langchain_ws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.1.0",
    "langchain-ollama>=1.0.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
]
